GameStrategy Add Armors - LetherArmor
