--- src/main/java/Examples/GameStrategy/Equipment/Armors/Gloves/Gloves.java
@@ -1,9 +1,0 @@
-package Examples.GameStrategy.Equipment.Armors.Gloves;
-
-import Examples.GameStrategy.Equipment.Armors.Cloths;
-
-public interface Gloves extends Cloths {
-
-	void wearGloves();
-
-}
