--- src/main/java/Examples/GameStrategy/Equipment/Armors/Helmets/Helmet.java
@@ -1,7 +1,0 @@
-package Examples.GameStrategy.Equipment.Armors.Helmets;
-
-import Examples.GameStrategy.Equipment.Armors.Cloths;
-
-public interface Helmet extends Cloths {
-	void wearHelmet();
-}
