+++ src/main/java/Examples/GameStrategy/Equipment/Cloths/Armors/Armor.java
@@ -1,0 +1,7 @@
+package Examples.GameStrategy.Equipment.Cloths.Armors;
+
+import Examples.GameStrategy.Equipment.Cloths.Cloths;
+
+public interface Armor extends Cloths {
+	void wearArmor();
+}
