+++ src/main/java/Examples/GameStrategy/Equipment/Cloths/Gloves/Gloves.java
@@ -1,0 +1,9 @@
+package Examples.GameStrategy.Equipment.Cloths.Gloves;
+
+import Examples.GameStrategy.Equipment.Cloths.Cloths;
+
+public interface Gloves extends Cloths {
+
+	void wearGloves();
+
+}
