+++ src/main/java/Examples/GameStrategy/Equipment/Cloths/Helmets/Helmet.java
@@ -1,0 +1,7 @@
+package Examples.GameStrategy.Equipment.Cloths.Helmets;
+
+import Examples.GameStrategy.Equipment.Cloths.Cloths;
+
+public interface Helmet extends Cloths {
+	void wearHelmet();
+}
